move invert to shared
