+++ packages/server/src/index.ts
@@ -1,13 +1,13 @@
-// FIXME: this file should only export
-// - `initTRPC`
-// - `TRPCError`
-// - (maybe something else?)
-
-export * from './transformer';
-export * from './error/TRPCError';
-export * from './core';
-export { type DefaultErrorShape } from './error/formatter';
-export type { RootConfig, AnyRootConfig } from './core/internals/config';
+export { TRPCError } from './error/TRPCError';
+export {
+  initTRPC,
+  experimental_standaloneMiddleware,
+  type inferRouterInputs,
+  type inferRouterOutputs,
+  type inferProcedureInput,
+  type inferProcedureOutput,
+  type inferRouterError,
+} from './core';
 
 export type {
   /**
@@ -16,3 +16,11 @@ export type {
    */
   inferAsyncReturnType,
 } from './types';
+
+export type {
+  AnyRouter as AnyTRPCRouter,
+  /**
+   * @deprecated use `AnyTRPCRouter` instead
+   */
+  AnyRouter,
+} from './core';
